APT: adapt to ALT's lazy pkgCacheFile; PK needs this for really correct work
PackageKit expects that it is allowed to call BuildCaches() etc many
times, expecting that pkgCacheFile is lazy to build them.  If (like
before) BuildCaches() doesn't have an idempotent semantics, this leads
to incorrect work: old resources leak, saved old pointers are not
valid pointers to the objects currently owned by the pkgCacheFile
instance.

This was understood when investigating a crash in PackageKit that
happened after the leak of old pointers in APT's pkgCacheFile was
curtailed (without understanding that the owned pointers would better
not allowed to change at all during the lifespan of a pkgCacheFile
instance).