aptcc: mark as const arguments in the implementation of a method
...as a little help in maintaining the code. (The more things are
explicit and rigid, the easier it will be to change the code without
being afraid of introducing errors. For example, this method might
some day take arguments by reference.)

(Adding the const modifiers in the declaration in the header is just a
matter of taste/style. There is no need for them in the header, since
they don't affect the API, but just express a property of the
implementation of the function: the body of the function doesn't need
to modify these copies.)