apt: Don't crash while resolving mimetypes, modernize mime search code
Previously we warned when there was no AppStream metadata at all by
trying to fetch all metadata and warning if we got no result.
Given the amount of data Debian and Ubuntu ship with now, this is a bad
idea (as it slows down unsuccessful searches massively), and we should
do this again only when AsPool has gained a more efficient is_empty()
method. Therefore, this check was removed for now.

In addition, we stop monitoring the pool for changes, as we will
immediately close it anyway, so the additional overhead of setting up
background reloading is not needed. Same applies for loading Flatpak
data, which we are not concerned with in PackageKit.

We also validate the package-name C string now for being NULL before
converting it into a std::string, which resolves a crash.

(cherry picked from commit 80a708c59b7aa57c7ca58dddddb67767766fb077
v1.2.7~35)