(.spec) to avoid useless warnings, don't set a C++ std in optflags
The warnings were from the C compiler, which got a C++ option.

C++ std was needed because APT headers used new C++ features. (Setting
it here is a workaround instead of having it in APT pkg-config.) Now,
it's set in meson.build to a value which is large enough for APT
headers and valid for the compilers in sisyphus, p10, sisyphus_e2k,
p10_e2k. Since PACKAGEKIT_1_2_5~4
25de1eba8 (trivial: aptcc: Build with older Meson versions, 2022-02-17).

If one needs to set it again by force, it'd be nicer to pass it as
-Dcpp_std=c++17 to meson or as part of apt_dep in meson.build
(emulating a pkg-config for APT, which itself would be an even better
solution).

(BTW, the current e2k -Wno-error flag doesn't deprive us of the
interesting fatal compiler errors there as long as individual
-Werror=* are used (in meson.build) instead of -Werror.  Otherwise,
one could put %%remove_optflags -Wno-error in the spec-file.)

(cherry picked from commit 9890681beac9615ec30a941e304acecb15af4c77)