pk-console: make the formats for ErrorCode & ErrorDetails as intended (with \t)
The strings were printed, but there was no tab as intended. So we put
'\t' after ':' similarly to the other formats in this source file.

Noticed as compiler (lcc) warning #192: unrecognized character escape sequence

(This glitch goes way back to the first implementation of
pk_console_offline_status in PACKAGEKIT_0_8_10~4
10a3ef17c (Add three pkcon sub-commands for offline updates, 2013-07-26).)